+++ src/components/token/token.dal.js
@@ -2,6 +2,7 @@ import _pickBy from 'lodash.pickby';
 
 import TokenDTO from './token.dto';
 import TokenModel from './token.model';
+import { InstanceofError } from '@libraries/Error';
 
 export const getOneByUserId = async userIdParameter => {
   try {
@@ -23,6 +24,9 @@ export const getOneByToken = async tokenParameter => {
 
 export const create = async tokenDTOParameter => {
   try {
+    if (!(tokenDTOParameter instanceof TokenDTO))
+      throw new InstanceofError('Param sent need to be an TokenDTO.');
+
     const tokenDTO = _pickBy(tokenDTOParameter);
     const tokenModel = new TokenModel(tokenDTO);
     const token = await tokenModel.save();
@@ -34,6 +38,9 @@ export const create = async tokenDTOParameter => {
 
 export const update = async tokenDTOParameter => {
   try {
+    if (!(tokenDTOParameter instanceof TokenDTO))
+      throw new InstanceofError('Param sent need to be an TokenDTO.');
+
     const tokenDTO = _pickBy(tokenDTOParameter);
     const token = await TokenModel.findOneAndUpdate({ _id: tokenDTO.id }, tokenDTO, { new: true });
     return new TokenDTO(token);
@@ -44,6 +51,9 @@ export const update = async tokenDTOParameter => {
 
 export const remove = async tokenDTOParameter => {
   try {
+    if (!(tokenDTOParameter instanceof TokenDTO))
+      throw new InstanceofError('Param sent need to be an TokenDTO.');
+
     await TokenModel.findOneAndRemove({ _id: tokenDTOParameter.id });
   } catch (err) {
     throw err;
